Note page is added and some other changes are made.
