+++ index.html
@@ -14,6 +14,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/JavaScript_Files/main.tsx"></script>
+    <script type="module" src="/src/Components/main.tsx"></script>
   </body>
 </html>
