+++ src/CSS_Files/_Variables.scss
@@ -1,4 +1,5 @@
 $mainColor: #9fecd0;
 $mainColorLight: #d4f7ea;
+$activeButtonColorLight: #e4fff5;
 $cardBorderColor: #cccccc;
 $elementColorOnNoNote: #dddcdc;
