+++ src/Components/main.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 
 // Component imports
-import App from "../Components/App";
+import App from "./App";
 
 // CSS imports
 import "../CSS_Files/style.scss";
